requirements.txt
streamlit
langchain-google-genai
langchain
faiss-cpu
pypdf
python-dotenv
nest-asyncio